# docs/openapi.yml
openapi: 3.0.2
info:
  title: 'b2b-marketplace'
  description: 'API веб-сервиса для взаимодействия между продавцами и покупателями оптовых товаров'
  version: '0.0.1'
paths:
...
  /api/v1/baskets/mine:
    post:
      summary: 'Добавить товар в корзину'
      tags:
        - Корзины
      security:
        - Token: []
      operationId: 'Добавление товара в корзину'
      description: 'Доступно только авторизированным пользователям'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasketAddProduct'
      responses:
        '201':
          description: 'Товар успешно добавлен'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: 'Обновить товары в корзине'
      tags:
        - Корзины
      security:
        - Token: []
      operationId: 'Обновление товаров в корзине пользователя'
      description: 'Доступно только авторизированным пользователям'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasketUpdateProducts'
      responses:
        '201':
          description: 'Товар успешно добавлен'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      summary: 'Просмотреть корзину пользователя'
      tags:
        - Корзины
      security:
        - Token: []
      operationId: 'Получение информации о корзине пользователя'
      description: 'Доступно только авторизированному пользователю'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basket'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Category:
      description: 'Категория товара'
      type: object
      properties:
        id:
          type: integer          
          example: 10
          description: 'Уникальный идентификатор категории'
        name:
          type: string
          example: Куртки
          description: 'Наименование категории'
        slug:
          type: string
          example: kurtki
          description: 'Часть url категории (слаг)'
        parent_id:
          type: integer
          example: 123
          description: 'Уникальный идентификатор родительской категории'
    ProductsList:
      type: object
      properties:
        id:
          type: integer
          description: 'Уникальный id'
        sku:
          type: string
          description: 'артикул товара'
        name:
          type: string
          description: 'название товара'
        brand:
          type: string
          description: 'производитель товара'
        price:
          type: number
          description: 'цена товара'
        wholesale_quantity:
          type: integer
          description: 'минимальный объем заказа'
        video:
          type: string
          description: 'видео'
        quantity_in_stock:
          type: integer
          description: 'количество в наличии'
        description:
          type: string
          description: 'описание'
        manufacturer_country:
          type: string
          description: 'страна производитель'
        user:
          type: object
          items:
            example:
              id: 1
              username: 'string'
            type: object
          properties:
            id:
              type: integer
              description: 'id пользователя'
            username:
              type: string
              description: 'username'
        category:
          $ref: '#/components/schemas/Category'
        images:
          type: array
          description: 'список изображений'
          items:
            example:
              image: 'http://example.org/images/image.jpeg'
            type: object
            properties:
              image:
                description: 'Ссылка на картинку на сайте'
                example: 'http://example.org/images/image.jpeg'
                type: string
                format: url
    Basket:
      description: 'Корзина'
      type: object
      properties:
        id:
          type: integer
          example: 3
          description: 'Уникальный идентификатор корзины'
        products:
          type: array
          description: 'Список товаров в корзине'
          items:
            properties:
              product:
                $ref: '#/components/schemas/ProductsList'
              quantity:
                type: integer
                example: 1000
                description: 'Количество товаров'
    BasketAddProduct:
      description: 'Тело запроса для добавления товара в корзину'
      type: object
      properties:
        product:
          type: object
          properties:
            product_id:
              type: integer
              example: 4
              description: 'Уникальный идентификатор товара'
            quantity:
              type: integer
              example: 234
              description: 'Количество единиц товара'
    BasketUpdateProducts:
      description: 'Тело запроса для обновления товаров в корзине'
      type: array
      items:
        $ref: '#/components/schemas/BasketAddProduct'
...
    AuthenticationError:
      description: Пользователь не авторизован
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
    PermissionDenied:
      description: Недостаточно прав
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "У вас недостаточно прав для выполнения данного действия."
          type: string
    NotFound:
      description: Объект не найден
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Страница не найдена."
          type: string
...
  responses:
...
    AuthenticationError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
    PermissionDenied:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
  securitySchemes:
    Token:
      type: http
      scheme: bearer
      bearerFormat: JWT